JumpPointsExpansionPolicy:  - when expanding nodes with no parent (i.e. start node) 	we now compute all jump points in every direction rather than 	designating the immediate neighbours of the start node jump points.
Might roll this back later as it can negatively affect performance for small
searches (pointlessly explore an area for jump points when the goal is not in
that direction). Or, else, tweak it so we only identify jump points in any 
given direction if the direction is complementary to the goal direction as
indicated by the heuristic function.

hog: now avoids re-creating the reference map for each experiment.

JPAExpansionPolicy: messing around with order in which neighbours are
added to neighbour set.